requirements.txt
streamlit
pandas
plotly
paho-mqtt
python-dotenv